src/emoji/game/requirements.txt
flask
flask_session
pyopenssl
